# backend/src/SzApp.Api/openapi/SzApp.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SzApp.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/master-data/companies": {
      "post": {
        "tags": [
          "Master Data - Companies"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/contracts": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "parameters": [
          {
            "name": "Mode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "CompanyId",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "LocationCategoryId",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "IdSearch",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IsActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCompanyRequest": {
        "required": [
          "partnerId",
          "managerId",
          "shortName",
          "printName",
          "relativeFolderName",
          "companyTypeId",
          "vatTypeId",
          "ledgerEntryDate",
          "locationCategoryId",
          "note",
          "sortIndex",
          "externalAccount"
        ],
        "type": "object",
        "properties": {
          "partnerId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "managerId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "shortName": {
            "type": "string"
          },
          "printName": {
            "type": "string"
          },
          "relativeFolderName": {
            "type": [
              "null",
              "string"
            ]
          },
          "companyTypeId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "vatTypeId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ledgerEntryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "locationCategoryId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          },
          "sortIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "externalAccount": {
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}